ref(node): Split up incoming & outgoing http handling
Just splitting the code into multiple files to make it a bit easier to reason about things & update things in follow up PRs.

Also removing node-fetch types from node package (they are unused there?) and moving `http/index.ts` and `node-fetch/index.ts` to top level to avoid an otherwise empty folder.
