+++ packages/node-core/src/integrations/http/SentryHttpInstrumentation.ts
@@ -1,39 +1,21 @@
-/* eslint-disable max-lines */
 import type { ChannelListener } from 'node:diagnostics_channel';
 import { subscribe, unsubscribe } from 'node:diagnostics_channel';
 import type * as http from 'node:http';
 import type * as https from 'node:https';
-import type { EventEmitter } from 'node:stream';
-import { context, propagation } from '@opentelemetry/api';
+import { context } from '@opentelemetry/api';
 import { isTracingSuppressed } from '@opentelemetry/core';
 import type { InstrumentationConfig } from '@opentelemetry/instrumentation';
 import { InstrumentationBase, InstrumentationNodeModuleDefinition } from '@opentelemetry/instrumentation';
-import type { AggregationCounts, Client, SanitizedRequestData, Scope } from '@sentry/core';
-import {
-  addBreadcrumb,
-  addNonEnumerableProperty,
-  debug,
-  generateSpanId,
-  getBreadcrumbLogLevelFromHttpStatusCode,
-  getClient,
-  getCurrentScope,
-  getIsolationScope,
-  getSanitizedUrlString,
-  getTraceData,
-  httpRequestToRequestData,
-  isError,
-  LRUMap,
-  parseUrl,
-  SDK_VERSION,
-  stripUrlQueryAndFragment,
-  withIsolationScope,
-} from '@sentry/core';
-import { shouldPropagateTraceForUrl } from '@sentry/opentelemetry';
+import { debug, LRUMap, SDK_VERSION } from '@sentry/core';
 import { DEBUG_BUILD } from '../../debug-build';
-import { mergeBaggageHeaders } from '../../utils/baggage';
 import { getRequestUrl } from '../../utils/getRequestUrl';
-
-const INSTRUMENTATION_NAME = '@sentry/instrumentation-http';
+import { INSTRUMENTATION_NAME } from './constants';
+import { instrumentServer } from './incoming-requests';
+import {
+  addRequestBreadcrumb,
+  addTracePropagationHeadersToOutgoingRequest,
+  getRequestOptions,
+} from './outgoing-requests';
 
 type Http = typeof http;
 type Https = typeof https;
@@ -116,9 +98,6 @@ export type SentryHttpInstrumentationOptions = InstrumentationConfig & {
   sessionFlushingDelayMS?: number;
 };
 
-// We only want to capture request bodies up to 1mb.
-const MAX_BODY_BYTE_LENGTH = 1024 * 1024;
-
 /**
  * This custom HTTP instrumentation is used to isolate incoming requests and annotate them with additional information.
  * It does not emit any spans.
@@ -151,7 +130,12 @@ export class SentryHttpInstrumentation extends InstrumentationBase<SentryHttpIns
 
     const onHttpServerRequestStart = ((_data: unknown) => {
       const data = _data as { server: http.Server };
-      this._patchServerEmitOnce(data.server);
+      instrumentServer(data.server, {
+        ignoreIncomingRequestBody: this.getConfig().ignoreIncomingRequestBody,
+        maxIncomingRequestBodySize: this.getConfig().maxIncomingRequestBodySize,
+        trackIncomingRequestsAsSessions: this.getConfig().trackIncomingRequestsAsSessions,
+        sessionFlushingDelayMS: this.getConfig().sessionFlushingDelayMS ?? 60_000,
+      });
     }) satisfies ChannelListener;
 
     const onHttpClientResponseFinish = ((_data: unknown) => {
@@ -245,142 +229,7 @@ export class SentryHttpInstrumentation extends InstrumentationBase<SentryHttpIns
       return;
     }
 
-    // Add trace propagation headers
-    const url = getRequestUrl(request);
-
-    // Manually add the trace headers, if it applies
-    // Note: We do not use `propagation.inject()` here, because our propagator relies on an active span
-    // Which we do not have in this case
-    const tracePropagationTargets = getClient()?.getOptions().tracePropagationTargets;
-    const addedHeaders = shouldPropagateTraceForUrl(url, tracePropagationTargets, this._propagationDecisionMap)
-      ? getTraceData()
-      : undefined;
-
-    if (!addedHeaders) {
-      return;
-    }
-
-    const { 'sentry-trace': sentryTrace, baggage } = addedHeaders;
-
-    // We do not want to overwrite existing header here, if it was already set
-    if (sentryTrace && !request.getHeader('sentry-trace')) {
-      try {
-        request.setHeader('sentry-trace', sentryTrace);
-        DEBUG_BUILD && debug.log(INSTRUMENTATION_NAME, 'Added sentry-trace header to outgoing request');
-      } catch (error) {
-        DEBUG_BUILD &&
-          debug.error(
-            INSTRUMENTATION_NAME,
-            'Failed to add sentry-trace header to outgoing request:',
-            isError(error) ? error.message : 'Unknown error',
-          );
-      }
-    }
-
-    if (baggage) {
-      // For baggage, we make sure to merge this into a possibly existing header
-      const newBaggage = mergeBaggageHeaders(request.getHeader('baggage'), baggage);
-      if (newBaggage) {
-        try {
-          request.setHeader('baggage', newBaggage);
-          DEBUG_BUILD && debug.log(INSTRUMENTATION_NAME, 'Added baggage header to outgoing request');
-        } catch (error) {
-          DEBUG_BUILD &&
-            debug.error(
-              INSTRUMENTATION_NAME,
-              'Failed to add baggage header to outgoing request:',
-              isError(error) ? error.message : 'Unknown error',
-            );
-        }
-      }
-    }
-  }
-
-  /**
-   * Patch a server.emit function to handle process isolation for incoming requests.
-   * This will only patch the emit function if it was not already patched.
-   */
-  private _patchServerEmitOnce(server: http.Server): void {
-    // eslint-disable-next-line @typescript-eslint/unbound-method
-    const originalEmit = server.emit;
-
-    // This means it was already patched, do nothing
-    if ((originalEmit as { __sentry_patched__?: boolean }).__sentry_patched__) {
-      return;
-    }
-
-    DEBUG_BUILD && debug.log(INSTRUMENTATION_NAME, 'Patching server.emit');
-
-    // eslint-disable-next-line @typescript-eslint/no-this-alias
-    const instrumentation = this;
-    const { ignoreIncomingRequestBody, maxIncomingRequestBodySize = 'medium' } = instrumentation.getConfig();
-
-    const newEmit = new Proxy(originalEmit, {
-      apply(target, thisArg, args: [event: string, ...args: unknown[]]) {
-        // Only traces request events
-        if (args[0] !== 'request') {
-          return target.apply(thisArg, args);
-        }
-
-        DEBUG_BUILD && debug.log(INSTRUMENTATION_NAME, 'Handling incoming request');
-
-        const isolationScope = getIsolationScope().clone();
-        const request = args[1] as http.IncomingMessage;
-        const response = args[2] as http.OutgoingMessage;
-
-        const normalizedRequest = httpRequestToRequestData(request);
-
-        // request.ip is non-standard but some frameworks set this
-        const ipAddress = (request as { ip?: string }).ip || request.socket?.remoteAddress;
-
-        const url = request.url || '/';
-        if (!ignoreIncomingRequestBody?.(url, request) && maxIncomingRequestBodySize !== 'none') {
-          patchRequestToCaptureBody(request, isolationScope, maxIncomingRequestBodySize);
-        }
-
-        // Update the isolation scope, isolate this request
-        isolationScope.setSDKProcessingMetadata({ normalizedRequest, ipAddress });
-
-        // attempt to update the scope's `transactionName` based on the request URL
-        // Ideally, framework instrumentations coming after the HttpInstrumentation
-        // update the transactionName once we get a parameterized route.
-        const httpMethod = (request.method || 'GET').toUpperCase();
-        const httpTarget = stripUrlQueryAndFragment(url);
-
-        const bestEffortTransactionName = `${httpMethod} ${httpTarget}`;
-
-        isolationScope.setTransactionName(bestEffortTransactionName);
-
-        if (instrumentation.getConfig().trackIncomingRequestsAsSessions !== false) {
-          recordRequestSession({
-            requestIsolationScope: isolationScope,
-            response,
-            sessionFlushingDelayMS: instrumentation.getConfig().sessionFlushingDelayMS ?? 60_000,
-          });
-        }
-
-        return withIsolationScope(isolationScope, () => {
-          // Set a new propagationSpanId for this request
-          // We rely on the fact that `withIsolationScope()` will implicitly also fork the current scope
-          // This way we can save an "unnecessary" `withScope()` invocation
-          getCurrentScope().getPropagationContext().propagationSpanId = generateSpanId();
-
-          // If we don't want to extract the trace from the header, we can skip this
-          if (!instrumentation.getConfig().extractIncomingTraceFromHeader) {
-            return target.apply(thisArg, args);
-          }
-
-          const ctx = propagation.extract(context.active(), normalizedRequest.headers);
-          return context.with(ctx, () => {
-            return target.apply(thisArg, args);
-          });
-        });
-      },
-    });
-
-    addNonEnumerableProperty(newEmit, '__sentry_patched__', true);
-
-    server.emit = newEmit;
+    addTracePropagationHeadersToOutgoingRequest(request, this._propagationDecisionMap);
   }
 
   /**
@@ -402,262 +251,3 @@ export class SentryHttpInstrumentation extends InstrumentationBase<SentryHttpIns
     return ignoreOutgoingRequests(url, options);
   }
 }
-
-/** Add a breadcrumb for outgoing requests. */
-function addRequestBreadcrumb(request: http.ClientRequest, response: http.IncomingMessage | undefined): void {
-  const data = getBreadcrumbData(request);
-
-  const statusCode = response?.statusCode;
-  const level = getBreadcrumbLogLevelFromHttpStatusCode(statusCode);
-
-  addBreadcrumb(
-    {
-      category: 'http',
-      data: {
-        status_code: statusCode,
-        ...data,
-      },
-      type: 'http',
-      level,
-    },
-    {
-      event: 'response',
-      request,
-      response,
-    },
-  );
-}
-
-function getBreadcrumbData(request: http.ClientRequest): Partial<SanitizedRequestData> {
-  try {
-    // `request.host` does not contain the port, but the host header does
-    const host = request.getHeader('host') || request.host;
-    const url = new URL(request.path, `${request.protocol}//${host}`);
-    const parsedUrl = parseUrl(url.toString());
-
-    const data: Partial<SanitizedRequestData> = {
-      url: getSanitizedUrlString(parsedUrl),
-      'http.method': request.method || 'GET',
-    };
-
-    if (parsedUrl.search) {
-      data['http.query'] = parsedUrl.search;
-    }
-    if (parsedUrl.hash) {
-      data['http.fragment'] = parsedUrl.hash;
-    }
-
-    return data;
-  } catch {
-    return {};
-  }
-}
-
-/**
- * This method patches the request object to capture the body.
- * Instead of actually consuming the streamed body ourselves, which has potential side effects,
- * we monkey patch `req.on('data')` to intercept the body chunks.
- * This way, we only read the body if the user also consumes the body, ensuring we do not change any behavior in unexpected ways.
- */
-function patchRequestToCaptureBody(
-  req: http.IncomingMessage,
-  isolationScope: Scope,
-  maxIncomingRequestBodySize: 'small' | 'medium' | 'always',
-): void {
-  let bodyByteLength = 0;
-  const chunks: Buffer[] = [];
-
-  DEBUG_BUILD && debug.log(INSTRUMENTATION_NAME, 'Patching request.on');
-
-  /**
-   * We need to keep track of the original callbacks, in order to be able to remove listeners again.
-   * Since `off` depends on having the exact same function reference passed in, we need to be able to map
-   * original listeners to our wrapped ones.
-   */
-  const callbackMap = new WeakMap();
-
-  const maxBodySize =
-    maxIncomingRequestBodySize === 'small'
-      ? 1_000
-      : maxIncomingRequestBodySize === 'medium'
-        ? 10_000
-        : MAX_BODY_BYTE_LENGTH;
-
-  try {
-    // eslint-disable-next-line @typescript-eslint/unbound-method
-    req.on = new Proxy(req.on, {
-      apply: (target, thisArg, args: Parameters<typeof req.on>) => {
-        const [event, listener, ...restArgs] = args;
-
-        if (event === 'data') {
-          DEBUG_BUILD &&
-            debug.log(INSTRUMENTATION_NAME, `Handling request.on("data") with maximum body size of ${maxBodySize}b`);
-
-          const callback = new Proxy(listener, {
-            apply: (target, thisArg, args: Parameters<typeof listener>) => {
-              try {
-                const chunk = args[0] as Buffer | string;
-                const bufferifiedChunk = Buffer.from(chunk);
-
-                if (bodyByteLength < maxBodySize) {
-                  chunks.push(bufferifiedChunk);
-                  bodyByteLength += bufferifiedChunk.byteLength;
-                } else if (DEBUG_BUILD) {
-                  debug.log(
-                    INSTRUMENTATION_NAME,
-                    `Dropping request body chunk because maximum body length of ${maxBodySize}b is exceeded.`,
-                  );
-                }
-              } catch (err) {
-                DEBUG_BUILD && debug.error(INSTRUMENTATION_NAME, 'Encountered error while storing body chunk.');
-              }
-
-              return Reflect.apply(target, thisArg, args);
-            },
-          });
-
-          callbackMap.set(listener, callback);
-
-          return Reflect.apply(target, thisArg, [event, callback, ...restArgs]);
-        }
-
-        return Reflect.apply(target, thisArg, args);
-      },
-    });
-
-    // Ensure we also remove callbacks correctly
-    // eslint-disable-next-line @typescript-eslint/unbound-method
-    req.off = new Proxy(req.off, {
-      apply: (target, thisArg, args: Parameters<typeof req.off>) => {
-        const [, listener] = args;
-
-        const callback = callbackMap.get(listener);
-        if (callback) {
-          callbackMap.delete(listener);
-
-          const modifiedArgs = args.slice();
-          modifiedArgs[1] = callback;
-          return Reflect.apply(target, thisArg, modifiedArgs);
-        }
-
-        return Reflect.apply(target, thisArg, args);
-      },
-    });
-
-    req.on('end', () => {
-      try {
-        const body = Buffer.concat(chunks).toString('utf-8');
-        if (body) {
-          // Using Buffer.byteLength here, because the body may contain characters that are not 1 byte long
-          const bodyByteLength = Buffer.byteLength(body, 'utf-8');
-          const truncatedBody =
-            bodyByteLength > maxBodySize
-              ? `${Buffer.from(body)
-                  .subarray(0, maxBodySize - 3)
-                  .toString('utf-8')}...`
-              : body;
-
-          isolationScope.setSDKProcessingMetadata({ normalizedRequest: { data: truncatedBody } });
-        }
-      } catch (error) {
-        if (DEBUG_BUILD) {
-          debug.error(INSTRUMENTATION_NAME, 'Error building captured request body', error);
-        }
-      }
-    });
-  } catch (error) {
-    if (DEBUG_BUILD) {
-      debug.error(INSTRUMENTATION_NAME, 'Error patching request to capture body', error);
-    }
-  }
-}
-
-function getRequestOptions(request: http.ClientRequest): http.RequestOptions {
-  return {
-    method: request.method,
-    protocol: request.protocol,
-    host: request.host,
-    hostname: request.host,
-    path: request.path,
-    headers: request.getHeaders(),
-  };
-}
-
-/**
- * Starts a session and tracks it in the context of a given isolation scope.
- * When the passed response is finished, the session is put into a task and is
- * aggregated with other sessions that may happen in a certain time window
- * (sessionFlushingDelayMs).
- *
- * The sessions are always aggregated by the client that is on the current scope
- * at the time of ending the response (if there is one).
- */
-// Exported for unit tests
-export function recordRequestSession({
-  requestIsolationScope,
-  response,
-  sessionFlushingDelayMS,
-}: {
-  requestIsolationScope: Scope;
-  response: EventEmitter;
-  sessionFlushingDelayMS?: number;
-}): void {
-  requestIsolationScope.setSDKProcessingMetadata({
-    requestSession: { status: 'ok' },
-  });
-  response.once('close', () => {
-    // We need to grab the client off the current scope instead of the isolation scope because the isolation scope doesn't hold any client out of the box.
-    const client = getClient();
-    const requestSession = requestIsolationScope.getScopeData().sdkProcessingMetadata.requestSession;
-
-    if (client && requestSession) {
-      DEBUG_BUILD && debug.log(`Recorded request session with status: ${requestSession.status}`);
-
-      const roundedDate = new Date();
-      roundedDate.setSeconds(0, 0);
-      const dateBucketKey = roundedDate.toISOString();
-
-      const existingClientAggregate = clientToRequestSessionAggregatesMap.get(client);
-      const bucket = existingClientAggregate?.[dateBucketKey] || { exited: 0, crashed: 0, errored: 0 };
-      bucket[({ ok: 'exited', crashed: 'crashed', errored: 'errored' } as const)[requestSession.status]]++;
-
-      if (existingClientAggregate) {
-        existingClientAggregate[dateBucketKey] = bucket;
-      } else {
-        DEBUG_BUILD && debug.log('Opened new request session aggregate.');
-        const newClientAggregate = { [dateBucketKey]: bucket };
-        clientToRequestSessionAggregatesMap.set(client, newClientAggregate);
-
-        const flushPendingClientAggregates = (): void => {
-          clearTimeout(timeout);
-          unregisterClientFlushHook();
-          clientToRequestSessionAggregatesMap.delete(client);
-
-          const aggregatePayload: AggregationCounts[] = Object.entries(newClientAggregate).map(
-            ([timestamp, value]) => ({
-              started: timestamp,
-              exited: value.exited,
-              errored: value.errored,
-              crashed: value.crashed,
-            }),
-          );
-          client.sendSession({ aggregates: aggregatePayload });
-        };
-
-        const unregisterClientFlushHook = client.on('flush', () => {
-          DEBUG_BUILD && debug.log('Sending request session aggregate due to client flush');
-          flushPendingClientAggregates();
-        });
-        const timeout = setTimeout(() => {
-          DEBUG_BUILD && debug.log('Sending request session aggregate due to flushing schedule');
-          flushPendingClientAggregates();
-        }, sessionFlushingDelayMS).unref();
-      }
-    }
-  });
-}
-
-const clientToRequestSessionAggregatesMap = new Map<
-  Client,
-  { [timestampRoundedToSeconds: string]: { exited: number; crashed: number; errored: number } }
->();
